refactor: bytes methods moved from account utils to bytes utils
