+++ src/utils/bytes.ts
@@ -8,6 +8,7 @@
 import { Data, Utils } from '@ethersphere/bee-js'
 import { bytesToHex } from './hex'
 import { BeeArgumentError } from './error'
+import CryptoJS from 'crypto-js'
 
 export const SPAN_SIZE = 8
 
@@ -148,4 +149,18 @@ export function assertFlexBytes<Min extends number, Max extends number = Min>(
   max: Max,
 ): asserts b is Utils.FlexBytes<Min, Max> {
   return Utils.assertFlexBytes(b, min, max)
+}
+
+/**
+ * Converts bytes to CryptoJS WordArray
+ */
+export function bytesToWordArray(data: Uint8Array): CryptoJS.lib.WordArray {
+  return CryptoJS.enc.Hex.parse(bytesToHex(data))
+}
+
+/**
+ * Converts word array to bytes
+ */
+export function wordArrayToBytes(data: CryptoJS.lib.WordArray): Uint8Array {
+  return Utils.hexToBytes(CryptoJS.enc.Hex.stringify(data))
 }
